apps/ai-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
matplotlib
seaborn
numpy
plotly
langchain-groq
langchain-core
pymongo
scikit-learn
pandas
joblib
requests
python-dotenv
tavily-python
ngrok
